pyproject.toml
[project]
name = "crustdata-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.4.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
